feat: basic integration with discord
