--- sapiens_tools/tests/mod.rs
@@ -1,1 +1,0 @@
-mod e2e;
